Serve all StockNear data over HTTP, drop Playwright, upgrade dependencies
The speculation page broke when StockNear's options page started loading
empty for the Playwright scraper: the symbol lookup came back with no IV,
IV rank, max pain or expirations, and the page invented a Saturday
expiration to fill the gap. This removes the browser from the data path
entirely, then brings the dependency set current.

Symbol-level data from MCP. The options chain, expirations and strike
list now come from the StockNear MCP server (overview plus open interest
by strike and expiry). Lookup went from 10-30s to under half a second.
Strategy templates snap to listed strikes instead of rounding the price
(HITI at $2.69 produced an unlisted $3 put). Static files are served with
Cache-Control: no-cache so a rebuilt container's JS reaches browsers.

Contract quotes and history from the contract JSON API. The browser had
only been a cookie jar for quotes; stocknear_contract_api posts to
/api/options-contract-history with httpx and profile cookies when present.
Contract history no longer clicks through the Download menu for a CSV:
parse_history_json reads the same endpoint, derives dte, keeps the later
row for a repeated date, and refuses a payload whose expiration, strike or
type differ from the contract requested, replacing the page-URL guard.
Syncs run concurrently (4 contracts in 0.5s vs ~15s plus 2-4s each).
Batch quotes keep one slot per input so the UI can match legs by index.

Playwright removed. StockNearScraper, the persistent browser and its
startup prewarm, six page-scraping debug routes, the ProGated/AuthExpired/
DownloadTimeout errors and the Firefox image layers are gone; the image
shrank from 2.24 GB to 1.26 GB. Playwright had been the only thing
installing greenlet, which SQLAlchemy's async engine needs, so the
dependency is now declared as sqlalchemy[asyncio].

Dependencies upgraded to current releases, including FastAPI 0.141
(Starlette 1.7), pandas 3, yfinance 1.7, Bokeh 3.10, pytest 9 and
pytest-asyncio 1.4. Starlette 1.x requires TemplateResponse(request,
name, ctx). base.html hardcoded BokehJS 3.6.2, which left every chart
blank against Bokeh 3.10 documents; the script version now comes from the
installed package. requests is no longer used and was dropped; httpx2 is
added for Starlette's test client.

Python 3.13 in the image, requires-python and .python-version. mypy app
is clean (36 errors fixed, mostly a reused query variable in
update_position that typed positions as trades, implicit Optional
defaults, and Decimal sums without a Decimal start). The options-chain
CLI printed a dataclass repr instead of JSON; fixed.

+++ app/services/bs_math.py
@@ -119,7 +119,7 @@ def calculate_option_price(
     option_type: str,
     spot: float,
     strike: float,
-    days_to_expiry: int,
+    days_to_expiry: float,
     volatility: float = DEFAULT_VOLATILITY,
     risk_free_rate: float = DEFAULT_RISK_FREE_RATE
 ) -> float:
